+++ Form/Extension/ChoiceTypeExtension.php
@@ -56,12 +56,16 @@ class ChoiceTypeExtension extends AbstractTypeExtension
     }
 
     /**
-     * Returns the name of the type being extended.
-     *
-     * @return string The name of the type being extended
+     * {@inheritdoc}
      */
     public function getExtendedType()
     {
-        return 'choice';
+        /*
+         * NEXT_MAJOR: Remove when dropping Symfony <2.8 support. It should
+         * simply be return 'Symfony\Component\Form\Extension\Core\Type\ChoiceType';
+         */
+        return method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix')
+            ? 'Symfony\Component\Form\Extension\Core\Type\ChoiceType'
+            : 'choice';
     }
 }