+++ Menu/Provider/GroupMenuProvider.php
@@ -138,14 +138,18 @@ class GroupMenuProvider implements MenuProviderInterface
 
             if (false === $menuItem->hasChildren()) {
                 $menuItem->setDisplay(false);
+            } elseif (!empty($group['keep_open'])) {
+                $menuItem->setAttribute('class', 'keep-open');
+                $menuItem->setExtra('keep_open', $group['keep_open']);
             }
         } else {
             foreach ($group['items'] as $item) {
                 if (isset($item['admin']) && !empty($item['admin'])) {
                     $admin = $this->pool->getInstance($item['admin']);
 
-                    // skip menu item if no `list` url is available or user doesn't have the LIST access rights
+                    // Do not display group if no `list` url is available or user doesn't have the LIST access rights
                     if (!$admin->hasRoute('list') || !$admin->hasAccess('list')) {
+                        $menuItem->setDisplay(false);
                         continue;
                     }